requirements.txt
httpx
aiohttp
anyio
readability-lxml
trafilatura
beautifulsoup4
pdfplumber
pydantic
pydantic-settings
tenacity
python-slugify
loguru
tiktoken
openai
orjson
qdrant-client
faiss-cpu
fastapi
uvicorn
sse-starlette
python-dotenv
ragas
sentence-transformers
python-jose
slowapi
pytest
typer
langdetect
cryptography
cryptography
PyPDF2
python-docx
sqlalchemy
alembic
bcrypt
python-multipart
passlib
aiosqlite